[Feat/#22] 영수증 사진 받아 처리하는 로직 임시 구현
완전구현 아닌 형태만 잡아둔 상태

+++ Refrigerator/src/main/java/com/example/project/domain/expense/dto/ExpenseStatisticsResponse.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import com.example.project.domain.expense.domain.Category;
 
 public record ExpenseStatisticsResponse(
-	    long totalAmount,
-	    Map<Category, Integer> categorySum,
+		Long totalAmount,
+	    Map<Category, Long> categorySum,
 	    int month
 	) {}
